Move `BaseDocStore` into `config-server`.
...and thus, having removed the only class in `data-store`, remove that
module entirely.
