+++ local-modules/@bayou/config-server/BaseDocStore.js
