--- local-modules/@bayou/data-store/package.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": {
-    "@bayou/typecheck": "local",
-    "@bayou/util-common": "local"
-  }
-}
